Add support to compose Navigation
This commit is the solution for udacity exercise
Step.02-Exercise-Lets-Navigate-Already. This commit
re-writes all Fragments and added support for
navigation compose.
