+++ app/src/main/res/values/strings.xml
@@ -36,4 +36,5 @@
     <string name="congratulations">Congratulations!</string>
     <string name="share">Share</string>
     <string name="sharing_not_available">Sharing Not Available</string>
+    <string name="back_button">back navigation button</string>
 </resources>
